requirements.txt
Django==3.1.5
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
djangorestframework==3.12.2
psycopg2-binary==2.8.5  # https://github.com/psycopg/psycopg2
djangorestframework-simplejwt==4.4.0
djangorestframework-jwt==1.11.0

# Testing
# ------------------------------------------------------------------------------
mypy==0.770  # https://github.com/python/mypy
django-stubs==1.5.0  # https://github.com/typeddjango/django-stubs
pytest==5.4.3  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.3  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
isort==4.3.21  # https://github.com/timothycrosley/isort
flake8==3.8.2  # https://github.com/PyCQA/flake8
flake8-isort==3.0.0  # https://github.com/gforcada/flake8-isort
black==20.8b1 # https://github.com/ambv/black
pylint-django==2.0.15  # https://github.com/PyCQA/pylint-django
brunette==0.2.0  # https://github.com/odwyersoftware/brunette
coverage==5.1  # https://github.com/nedbat/coveragepy